# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and get JWT token",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorsResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                },
                "security": []
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user (invalidate JWT token)",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh JWT token",
                "operationId": "authRefresh",
                "responses": {
                    "200": {
                        "description": "Token refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "Authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "operationId": "productIndex",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/q"
                    },
                    {
                        "$ref": "#/components/parameters/sort"
                    },
                    {
                        "$ref": "#/components/parameters/price_from"
                    },
                    {
                        "$ref": "#/components/parameters/price_to"
                    },
                    {
                        "$ref": "#/components/parameters/category_id"
                    },
                    {
                        "$ref": "#/components/parameters/in_stock"
                    },
                    {
                        "$ref": "#/components/parameters/rating_from"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginationResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorsResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "description": "Credentials for JWT authentication",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "test@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "ErrorResponseSchema": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": " string",
                        "example": "Error occurred"
                    }
                },
                "type": "object"
            },
            "PaginationResponse": {
                "description": "Paginated response wrapper",
                "properties": {
                    "data": {
                        "description": "List of items",
                        "type": "array",
                        "items": {}
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string"
                                },
                                "active": {
                                    "type": "boolean"
                                },
                                "page": {
                                    "type": "integer",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "nullable": true
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "to": {
                        "type": "integer",
                        "example": 5
                    },
                    "total": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "TokenResponse": {
                "description": "JWT token response",
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "ValidationErrorsResponseSchema": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": " string",
                        "example": "The selected sort is invalid."
                    },
                    "errors": {
                        "type": " object",
                        "example": {
                            "sort": [
                                "The selected sort is invalid."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "description": " Category resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Смартфоны"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "description": " Product resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "IPhone 15 Pro Max"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "in_stock": {
                        "type": "boolean"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "User resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Test User"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-23T09:20:07.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "ErrorResponse": {
                "description": "Common error responce",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponseSchema"
                        }
                    }
                }
            },
            "ValidationErrorsResponse": {
                "description": "Validation errors",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorsResponseSchema"
                        }
                    }
                }
            }
        },
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "description": "Page number",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "Count of elements on page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "q": {
                "name": "q",
                "in": "query",
                "description": "Substing for searching",
                "required": false,
                "schema": {
                    "type": "string",
                    "maxLength": 255
                }
            },
            "sort": {
                "name": "sort",
                "in": "query",
                "description": "Type of sorting",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "price_asc",
                        "price_desc",
                        "rating_desc",
                        "newest"
                    ]
                }
            },
            "price_from": {
                "name": "price_from",
                "in": "query",
                "description": "Product price from",
                "required": false,
                "schema": {
                    "type": "number",
                    "format": "float",
                    "minimum": 1
                }
            },
            "price_to": {
                "name": "price_to",
                "in": "query",
                "description": "Product price to",
                "required": false,
                "schema": {
                    "type": "number",
                    "format": "float",
                    "minimum": 1
                }
            },
            "category_id": {
                "name": "category_id",
                "in": "query",
                "description": "Choosen category id",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "in_stock": {
                "name": "in_stock",
                "in": "query",
                "description": "If product in stock",
                "required": false,
                "schema": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            },
            "rating_from": {
                "name": "rating_from",
                "in": "query",
                "description": "Product rating from",
                "required": false,
                "schema": {
                    "type": "number",
                    "format": "float",
                    "minimum": 1
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}